requirements.txt
# Core dependencies for Telegram File-to-Link Bot
pyrogram==2.0.106
aiohttp==3.9.5
aiofiles==23.2.1

# Cryptography and security
tgcrypto==1.2.5
cryptography>=41.0.0,<46.0.0

# Environment and configuration
python-dotenv==1.0.1

# Async utilities
asyncio-throttle==1.0.2

# Logging and monitoring
colorlog==6.8.2

# HTTP client for additional requests
httpx==0.27.0

# Memory optimization
psutil==6.0.0

# Date and time utilities
python-dateutil==2.9.0

# JSON handling improvements
orjson==3.10.7

# Additional utilities
humanize==4.10.0
